+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed incorrect call signature for `options.callbacks.post_set_workspace`
 - Fixed incorrect fallback for `resolve_note`.
+- Fixed async with minimal functions copied from `vim._async`.
 
 ## [v3.13.0](https://github.com/obsidian-nvim/obsidian.nvim/releases/tag/v3.13.0) - 2025-07-28